Added doc theme, created doc skeleton, started adding docstrings
