+++ docs/tom_alerts/brokers.rst
@@ -1,0 +1,49 @@
+Brokers
+=======
+
+*****************
+Base Broker Class
+*****************
+
+.. automodule:: tom_alerts.alerts
+    :members:
+
+
+******
+ALeRCE
+******
+
+.. automodule:: tom_alerts.brokers.alerce
+    :members:
+
+
+*******
+ANTARES
+*******
+
+.. automodule:: tom_alerts.brokers.antares
+    :members:
+
+
+******
+Lasair
+******
+
+.. automodule:: tom_alerts.brokers.lasair
+    :members:
+
+
+****
+MARS
+****
+
+.. automodule:: tom_alerts.brokers.mars
+    :members:
+
+
+*****
+Scout
+*****
+
+.. automodule:: tom_alerts.brokers.scout
+    :members:
