+++ docs/tom_alerts/models.rst
@@ -1,0 +1,5 @@
+Models
+======
+
+.. automodule:: tom_alerts.models
+    :members:
