+++ docs/tom_alerts/views.rst
@@ -1,0 +1,5 @@
+Views
+=====
+
+.. automodule:: tom_alerts.views
+    :members:
