+++ docs/tom_catalogs/index.rst
@@ -1,10 +1,9 @@
-Targets
-=======
+Catalogs
+========
 
 .. toctree::
   :maxdepth: 2
 
   models
   views
-  templatetags
-  utils
+  harvesters
