+++ docs/tom_catalogs/models.rst
@@ -1,0 +1,5 @@
+Models
+======
+
+.. automodule:: tom_catalogs.models
+    :members:
