+++ docs/tom_catalogs/views.rst
@@ -1,0 +1,5 @@
+Views
+=====
+
+.. automodule:: tom_catalogs.views
+    :members:
