+++ docs/tom_common/index.rst
@@ -1,0 +1,5 @@
+Common Module
+=============
+
+.. toctree::
+  :maxdepth: 2
