+++ docs/tom_dataproducts/models.rst
@@ -1,0 +1,5 @@
+Models
+======
+
+.. automodule:: tom_dataproducts.models
+    :members:
