+++ docs/tom_dataproducts/templatetags.rst
@@ -1,0 +1,5 @@
+TemplateTags
+============
+
+.. automodule:: tom_dataproducts.templatetags.dataproduct_extras
+    :members:
