+++ docs/tom_dataproducts/views.rst
@@ -1,0 +1,5 @@
+Views
+=====
+
+.. automodule:: tom_dataproducts.views
+    :members:
