+++ docs/tom_observations/facilities.rst
@@ -1,0 +1,33 @@
+Facilities
+==========
+
+*******************
+Base Facility Class
+*******************
+
+.. automodule:: tom_observations.facility
+    :members:
+
+
+******
+Gemini
+******
+
+.. automodule:: tom_observations.facilities.gemini
+    :members:
+
+
+***********************
+Las Cumbres Observatory
+***********************
+
+.. automodule:: tom_observations.facilities.lco
+    :members:
+
+
+****
+SOAR
+****
+
+.. automodule:: tom_observations.facilities.soar
+    :members:
