+++ docs/tom_observations/models.rst
@@ -1,0 +1,5 @@
+Models
+======
+
+.. automodule:: tom_observations.models
+    :members:
