+++ docs/tom_observations/templatetags.rst
@@ -1,0 +1,5 @@
+TemplateTags
+============
+
+.. automodule:: tom_observations.templatetags.observation_extras
+    :members:
