+++ docs/tom_observations/utils.rst
@@ -1,0 +1,5 @@
+Utilities
+=========
+
+.. automodule:: tom_observations.utils
+    :members:
