+++ docs/tom_observations/views.rst
@@ -1,0 +1,5 @@
+Views
+=====
+
+.. automodule:: tom_observations.views
+    :members:
