+++ docs/tom_targets/templatetags.rst
@@ -1,5 +1,5 @@
-TemplateTags
-============
+Template Tags
+=============
 
 .. automodule:: tom_targets.templatetags.targets_extras
     :members:
