+++ docs/tom_targets/utils.rst
@@ -1,0 +1,5 @@
+Utilities
+=========
+
+.. automodule:: tom_targets.utils
+    :members:
